completely overhaul imports and directory structure
Main wins:
 - almost completely eliminated the background task
 - only the material-validation endpoint still needs it, so it is now
   defined within the material_validation sub-module.
 - removed all the coupling between methods and utilities used for
   different endpoints.

Collection endpoints:
- every endpoint has its own sub-module named accordingly
- every endpoint imports a get_... method from that module
- every endpoint has its own test
- every endpoint, module, test, test-resource are named
  consitently
- response models are moved to the respective get_... mehtods

General:
- moved constants to constant.py
- moved elastics search attributes to elastic sub-module
- eliminated unused and dead code
