+++ src/app/api/api.py
@@ -14,44 +14,37 @@ from starlette.status import (
     HTTP_422_UNPROCESSABLE_ENTITY,
 )
 
-from app.api.collections.models import (
-    PendingMaterialsResponse,
-    StatsResponse, CollectionValidationStats,
+from app.api.collections.material_validation import (
+    get_material_validation,
+    MaterialValidationResponse,
+    material_validation_cache,
 )
-from app.api.collections.collection_validation import (
-    collections_with_missing_properties,
-    overall_stats,
-)
-from app.api.analytics.storage import global_storage, global_store
-from app.api.collections.counts import (
-    AggregationMappings,
-    CollectionTreeCount,
-    oer_collection_counts,
-)
+from app.api.collections.collection_validation import get_collection_validation, CollectionValidationStats
+from app.api.collections.counts import AggregationMappings, CollectionTreeCount, get_counts
 from app.api.collections.material_counts import (
     CollectionMaterialCount,
-    get_collection_material_counts,
+    get_material_counts,
 )
 from app.api.collections.pending_materials import (
     LearningMaterial,
     MissingAttributeFilter,
     materials_filter_params,
-    search_materials_with_missing_attributes,
+    get_pending_materials,
 )
+from app.api.collections.statistics import get_statistics, StatsResponse
 from app.api.collections.tree import CollectionNode
 from app.api.collections.pending_collections import (
-    search_collections_with_missing_attributes,
+    get_pending_collections,
     MissingMaterials,
 )
-from app.api.collections.tree import build_collection_tree
+from app.api.collections.tree import get_tree
 from app.api.quality_matrix.collections import collection_quality_matrix
 from app.api.quality_matrix.models import Mode, QualityMatrix, Timeline
 from app.api.quality_matrix.replication_source import source_quality_matrix
 from app.api.quality_matrix.timeline import quality_backup, timestamps
 from app.api.collections.score import ScoreOutput, get_score
-from app.core.config import API_DEBUG
 from app.core.constants import COLLECTION_NAME_TO_ID, COLLECTION_ROOT_ID
-from app.core.models import ElasticResourceAttribute
+from app.elastic.attributes import ElasticResourceAttribute
 
 
 def get_database(request: Request) -> Database:
@@ -235,7 +228,7 @@ async def get_timestamps(
     tags=["Collections"],
     summary="The average ratio of non-empty properties for the chosen collection",
 )
-async def score(*, node_id: uuid.UUID = Depends(node_ids_for_major_collections)):
+async def collection_score(*, node_id: uuid.UUID = Depends(node_ids_for_major_collections)):
     """
     Validate attributes of sub-collections and materials of the given collection grouped by OER and non-OER content.
 
@@ -304,7 +297,7 @@ async def ping_api():
     tags=["Collections"],
     summary="Provide the sub-tree of the collection hierarchy starting at given node",
 )
-async def get_collection_tree(*, node_id: uuid.UUID = Depends(node_ids_for_major_collections)):
+async def collection_tree(*, node_id: uuid.UUID = Depends(node_ids_for_major_collections)):
     """
     Returns a list of the immediate child nodes of the provided parent node (`node_id` path parameter).
 
@@ -314,7 +307,7 @@ async def get_collection_tree(*, node_id: uuid.UUID = Depends(node_ids_for_major
     **FIXME: See [Issue-86](https://github.com/openeduhub/metaqs-main/issues/86)**
     """
     validate_node_id(node_id)
-    return build_collection_tree(node_id).children
+    return get_tree(node_id).children
 
 
 @router.get(
@@ -325,7 +318,7 @@ async def get_collection_tree(*, node_id: uuid.UUID = Depends(node_ids_for_major
     responses={HTTP_404_NOT_FOUND: {"description": "Collection not found"}},
     tags=["Collections"],
 )
-async def get_collection_counts(
+async def collection_counts(
     *,
     node_id: uuid.UUID = Depends(node_ids_for_major_collections),
     facet: AggregationMappings = Param(
@@ -344,7 +337,7 @@ async def get_collection_counts(
     Within the elements of the list, the number of materials is grouped by OER vs Non-OER and the selected facet.
     """
     validate_node_id(node_id)
-    return await oer_collection_counts(node_id=node_id, facet=facet)
+    return await get_counts(node_id=node_id, facet=facet)
 
 
 @router.get(
@@ -355,7 +348,7 @@ async def get_collection_counts(
     responses={HTTP_404_NOT_FOUND: {"description": "Collection not found"}},
     tags=["Collections"],
 )
-async def filter_pending_collections(
+async def collection_pending_collections(
     *,
     node_id: uuid.UUID = Depends(node_ids_for_major_collections),
     missing_attribute: str = Path(
@@ -397,7 +390,7 @@ async def filter_pending_collections(
         missing_attribute = ElasticResourceAttribute.COLLECTION_DESCRIPTION
     else:
         raise HTTPException(status_code=400, detail=f"Invalid collection attribute: {missing_attribute}")
-    return await search_collections_with_missing_attributes(node_id, missing=missing_attribute)
+    return await get_pending_collections(node_id, missing=missing_attribute)
 
 
 @router.get(
@@ -408,7 +401,7 @@ async def filter_pending_collections(
     responses={HTTP_404_NOT_FOUND: {"description": "Collection not found"}},
     tags=["Collections"],
 )
-async def filter_materials_with_missing_attributes(
+async def collection_pending_materials(
     *,
     node_id: uuid.UUID = Depends(node_ids_for_major_collections),
     missing_attr_filter: MissingAttributeFilter = Depends(materials_filter_params),
@@ -448,7 +441,7 @@ async def filter_materials_with_missing_attributes(
     </b>
     """
     validate_node_id(node_id)
-    return await search_materials_with_missing_attributes(
+    return await get_pending_materials(
         collection_id=node_id,
         # fixme: resolve the whole attribute identification mess
         missing=getattr(ElasticResourceAttribute, str(missing_attr_filter.attr.name)),
@@ -468,8 +461,8 @@ async def collection_material_counts(*, node_id: uuid.UUID = Depends(node_ids_fo
     Returns the number of materials connected to all collections below this 'node_id' as a flat list.
     """
     validate_node_id(node_id)
-    collection = build_collection_tree(node_id=node_id)
-    return await get_collection_material_counts(collection=collection)
+    collection = get_tree(node_id=node_id)
+    return await get_material_counts(collection=collection)
 
 
 @router.get(
@@ -487,12 +480,9 @@ async def read_stats(
     Returns the number of materials found connected to the collection (`node_id` path parameter) and its
     sub-collections as well as materials containing the name of the respective collection, e.g., in the title.
     It is therefore an overview of materials, which could be added to a collection in the future.
-
-    This endpoint relies on an internal periodic background process which is scheduled to regularly update the data.
-    Its frequency can be configured via the BACKGROUND_TASK_TIME_INTERVAL environment variable.
     """
     validate_node_id(node_id)
-    return await overall_stats(node_id)
+    return await get_statistics(node_id)
 
 
 @router.get(
@@ -503,30 +493,27 @@ async def read_stats(
     responses={HTTP_404_NOT_FOUND: {"description": "Collection not found"}},
     tags=["Collections"],
 )
-async def read_stats_validation_collection(
+async def collection_collection_validation(
     *,
     node_id: uuid.UUID = Depends(node_ids_for_major_collections),
 ):
     """
     Returns the number of collections missing certain properties for this collection's 'node_id' and its
     sub-collections.
-
-    This endpoint relies on an internal periodic background process which is scheduled to regularly update the data.
-    Its frequency can be configured via the BACKGROUND_TASK_TIME_INTERVAL environment variable.
     """
     validate_node_id(node_id)
-    return collections_with_missing_properties(node_id)
+    return get_collection_validation(node_id)
 
 
 @router.get(
     "/collections/{node_id}/material-validation",
-    response_model=PendingMaterialsResponse,
+    response_model=MaterialValidationResponse,
     response_model_exclude_unset=True,
     status_code=HTTP_200_OK,
     responses={HTTP_404_NOT_FOUND: {"description": "Collection not found"}},
     tags=["Collections"],
 )
-async def read_material_validation(
+async def collection_material_validation(
     *,
     node_id: uuid.UUID = Depends(node_ids_for_major_collections),
 ):
@@ -542,19 +529,11 @@ async def read_material_validation(
     """
     validate_node_id(node_id)
     try:
-        return global_store.pending_materials[node_id]
+        if False:  # Fixme: if possible optimize the query and return real time data
+            return get_material_validation(collection_id=node_id)  # noqa
+        return material_validation_cache[node_id]
     except KeyError:
         raise HTTPException(
             status_code=503,
             detail=f"Background calculation for collection {node_id} incomplete. Please try again in a while.",
         )
-
-
-if API_DEBUG:
-
-    @router.get(
-        "/global",
-        description="""A debug endpoint to access the data stored inside the global storage.""",
-    )
-    async def get_global():
-        return global_storage
