+++ src/app/core/constants.py
@@ -29,3 +29,5 @@ COLLECTION_NAME_TO_ID: dict[str, str] = {
 
 OPEN_API_VERSION = "2.1.0"
 COLLECTION_ROOT_ID = "5e40e372-735c-4b17-bbf7-e827a5702b57"
+OER_LICENSES = ["CC_0", "PDM", "CC_BY", "CC_BY_SA"]
+FORBIDDEN_LICENSES = ["UNTERRICHTS_UND_LEHRMEDIEN", "NONE", ""]
