+++ tests/resources/material-counts-request.json
