+++ tests/resources/material-validation-single-collection-request.json
