+++ tests/resources/material-validation-single-collection-response.json
