+++ tests/resources/pending-collections-request.json
