--- tests/unit_tests/resources/test_global.json
@@ -1,86 +1,0 @@
-{
-  "collections": [
-    {
-      "id": "f3dc9ea1-d608-4b4e-a78c-98063a3e8461",
-      "doc": {
-        "path": [
-          "5e40e372-735c-4b17-bbf7-e827a5702b57",
-          "11bdb8a0-a9f5-4028-becc-cbf8e328dd4b"
-        ],
-        "nodeRef": {
-          "storeRef": {
-            "identifier": "SpacesStore",
-            "protocol": "workspace"
-          },
-          "id": "f3dc9ea1-d608-4b4e-a78c-98063a3e8461"
-        },
-        "properties": {
-          "cclom:title": "dummy_title"
-        }
-      },
-      "derived_at": "2022-07-05T16:50:18.789156"
-    }
-  ],
-  "counts": [
-    {
-      "noderef_id": "f3dc9ea1-d608-4b4e-a78c-98063a3e8461",
-      "total": 18025,
-      "counts": {
-        "test_lrt": 3
-      }
-    }
-  ],
-  "counts_oer": [
-    {
-      "noderef_id": "f3dc9ea1-d608-4b4e-a78c-98063a3e8461",
-      "total": 18025,
-      "counts": {
-        "test_lrt": 3
-      }
-    }
-  ],
-  "materials": [
-    {
-      "id": "263afc5b-6445-4a5a-b014-a77f1db473b9",
-      "doc": {
-        "collections": [
-          {
-            "nodeRef": {
-              "id": "f3dc9ea1-d608-4b4e-a78c-98063a3e8461"
-            }
-          }
-        ],
-        "nodeRef": {
-          "storeRef": {
-            "identifier": "SpacesStore",
-            "protocol": "workspace"
-          },
-          "id": "263afc5b-6445-4a5a-b014-a77f1db473b9"
-        },
-        "properties": {
-          "ccm:objecttype": "dummy_object_type"
-        }
-      },
-      "derived_at": "2022-07-07T08:51:25.636346"
-    }
-  ],
-  "search": [
-    {
-      "node_id": "11bdb8a0-a9f5-4028-becc-cbf8e328dd4b",
-      "collections": {
-        "f3dc9ea1-d608-4b4e-a78c-98063a3e8461": {
-          "node_id": "f3dc9ea1-d608-4b4e-a78c-98063a3e8461",
-          "missing_materials": {
-            "http://w3id.org/openeduhub/vocabs/new_lrt_aggregated/3927fdb6-0477-422c-9f5a-6285948aeaf4": 6,
-            "http://w3id.org/openeduhub/vocabs/new_lrt_aggregated/05aa0f49-7e1b-498b-a7d5-c5fc8e73b2e2": 4,
-            "http://w3id.org/openeduhub/vocabs/new_lrt_aggregated/37a3ad9c-727f-4b74-bbab-27d59015c695": 4,
-            "http://w3id.org/openeduhub/vocabs/new_lrt_aggregated/38774279-af36-4ec2-8e70-811d5a51a6a1": 4,
-            "http://w3id.org/openeduhub/vocabs/new_lrt_aggregated/f1341358-3f91-449b-b6eb-f58636f756a0": 4,
-            "N/A": 1,
-            "total": 23
-          }
-        }
-      }
-    }
-  ]
-}
